style(css): reformat CSS for better readability and maintainability Reduce border width for images to 0px for a cleaner look
The CSS file is reformatted to improve readability by placing each selector on a new line. This change enhances maintainability by making it easier to identify and modify individual selectors. The border width for images is reduced to 0px to achieve a cleaner visual appearance.

feat(logo.html): add AHA! logo image to the logo section
Replace text-based logo with an image for better branding

The logo section now includes an image of the AHA! logo, replacing the previous text-based logo. This change enhances the visual branding of the site by using a consistent and recognizable logo image. The text-based logo code is commented out for potential future use.